requirements/llm.txt
torchtune
wandb
transformers
peft
quanto
accelerate
bitsandbytes
langchain-huggingface
